feat: split TopologyMetrics into TopologyMetrics & TopologyManagerMetrics
(cherry picked from commit 55f7d512c5a1a5251b6e0e28c6a95acacf6fdeaf)
