+++ zeebe/topology/src/main/java/io/camunda/zeebe/topology/metrics/TopologyMetrics.java
@@ -13,18 +13,12 @@ import io.camunda.zeebe.topology.state.ClusterChangePlan;
 import io.camunda.zeebe.topology.state.ClusterChangePlan.Status;
 import io.camunda.zeebe.topology.state.ClusterTopology;
 import io.camunda.zeebe.topology.state.CompletedChange;
-import io.camunda.zeebe.topology.state.TopologyChangeOperation;
 import io.camunda.zeebe.util.CloseableSilently;
 import io.camunda.zeebe.util.collection.Table;
 import io.camunda.zeebe.util.micrometer.EnumMeter;
-import io.camunda.zeebe.util.micrometer.MicrometerUtil;
-import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
-import io.micrometer.core.instrument.Timer;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 
 public final class TopologyMetrics {
@@ -35,8 +29,6 @@ public final class TopologyMetrics {
   private final AtomicLong changeVersion;
   private final AtomicLong pendingOperations;
   private final AtomicLong completedOperations;
-  private final Map<String, Timer> operationDuration;
-  private final Table<String, String, Counter> operationAttempts;
   private final EnumMeter<Status> changeStatus;
 
   public TopologyMetrics(final MeterRegistry registry) {
@@ -52,27 +44,6 @@ public final class TopologyMetrics {
     changeVersion = makeGauge(CHANGE_VERSION);
     pendingOperations = makeGauge(PENDING_OPERATIONS);
     completedOperations = makeGauge(COMPLETED_OPERATIONS);
-    operationDuration = new HashMap<>();
-    operationAttempts = Table.simple();
-  }
-
-  private Timer registerOperation(final String operation) {
-    return Timer.builder(OPERATION_DURATION.getName())
-        .description(OPERATION_DURATION.getDescription())
-        .serviceLevelObjectives(OPERATION_DURATION.getTimerSLOs())
-        .tags(TopologyMetricsKeyName.OPERATION.asString(), operation)
-        .register(registry);
-  }
-
-  private Counter registerAttempt(final String operation, final String outcome) {
-    return Counter.builder(OPERATION_ATTEMPTS.getName())
-        .description(OPERATION_DURATION.getDescription())
-        .tags(
-            TopologyMetricsKeyName.OPERATION.asString(),
-            operation,
-            TopologyMetricsKeyName.OUTCOME.asString(),
-            outcome)
-        .register(registry);
   }
 
   private AtomicLong makeGauge(final TopologyMetricsDoc meter) {
@@ -105,47 +76,5 @@ public final class TopologyMetrics {
             .map(ClusterChangePlan::completedOperations)
             .map(List::size)
             .orElse(0));
-  }
-
-  public OperationObserver observeOperation(final TopologyChangeOperation operation) {
-    return startOperation(operation);
-  }
-
-  OperationObserver startOperation(final TopologyChangeOperation operation) {
-    final var timer =
-        operationDuration.computeIfAbsent(
-            operation.getClass().getSimpleName(), this::registerOperation);
-    return new OperationObserver(operation, MicrometerUtil.timer(timer, Timer.start(registry)));
-  }
-
-  public final class OperationObserver {
-    private final TopologyChangeOperation operation;
-    private final CloseableSilently timer;
-
-    private OperationObserver(
-        final TopologyChangeOperation operation, final CloseableSilently timer) {
-      this.operation = operation;
-      this.timer = timer;
-    }
-
-    public void failed() {
-      timer.close();
-      operationAttempts
-          .computeIfAbsent(
-              operation.getClass().getSimpleName(),
-              Outcome.FAILED.getName(),
-              TopologyMetrics.this::registerAttempt)
-          .increment();
-    }
-
-    public void applied() {
-      timer.close();
-      operationAttempts
-          .computeIfAbsent(
-              operation.getClass().getSimpleName(),
-              Outcome.APPLIED.getName(),
-              TopologyMetrics.this::registerAttempt)
-          .increment();
-    }
   }
 }
